feat(javascript): enable use with NPM
