+++ typescript-javascript/index.cjs
@@ -1,0 +1,3 @@
+const qrcodegen = require("./dist/qrcodegen.cjs").default;
+
+module.exports = qrcodegen;
