+++ typescript-javascript/index.js
@@ -1,0 +1,3 @@
+import qrcodegen from "./dist-esm/qrcodegen.js";
+
+export default qrcodegen;
