+++ typescript-javascript/qrcodegen-input-demo.html
@@ -147,8 +147,8 @@
 				</tbody>
 			</table>
 		</form>
-		<script type="application/javascript" src="qrcodegen.js"></script>
-		<script type="application/javascript" src="qrcodegen-input-demo.js"></script>
+		<script type="module" src="dist-demo/qrcodegen.js"></script>
+		<script type="module" src="dist-demo/qrcodegen-input-demo.js"></script>
 		
 		<hr>
 		<p>Copyright © Project Nayuki – <a href="https://www.nayuki.io/page/qr-code-generator-library">https://www.nayuki.io/page/qr-code-generator-library</a></p>
