+++ typescript-javascript/qrcodegen-output-demo.html
@@ -57,8 +57,8 @@
 			<p>(Are the JavaScript files missing?)</p>
 			<p>(The JavaScript code needs to be compiled from the TypeScript code.)</p>
 		</div>
-		<script type="application/javascript" src="qrcodegen.js"></script>
-		<script type="application/javascript" src="qrcodegen-output-demo.js"></script>
+		<script type="module" src="dist-demo/qrcodegen.js"></script>
+		<script type="module" src="dist-demo/qrcodegen-output-demo.js"></script>
 		<hr>
 		<p>Copyright © Project Nayuki – <a href="https://www.nayuki.io/page/qr-code-generator-library">https://www.nayuki.io/page/qr-code-generator-library</a></p>
 	</body>
