+++ typescript-javascript/tsconfig.es5.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.base",
+  "compilerOptions": {
+    "target": "ES5",
+    "outDir": "dist"
+  }
+}
