+++ typescript-javascript/tsconfig.es6.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.base",
+  "compilerOptions": {
+    "target": "ES2015",
+    "outDir": "dist-esm"
+  }
+}
